Moved data_linking to linking folder
