+++ src/nplinker/scoring/linking/data_linking.py
@@ -40,19 +40,19 @@ except ImportError:
         'Warning: plotting functionality will not be available (missing matplotlib and/or seaborn)'
     )
 
-from ..genomics import GCF
-from ..metabolomics import MolecularFamily
-from ..metabolomics import Spectrum
-from .linking.data_linking_functions import calc_correlation_matrix
-from .linking.data_linking_functions import calc_likelihood_matrix
-from .linking.data_linking_functions import link_prob
-from .linking.data_linking_functions import pair_prob_approx
-from .linking.data_linking_functions import pair_prob_hg
+from ...genomics import GCF
+from ...metabolomics import MolecularFamily
+from ...metabolomics import Spectrum
+from .data_linking_functions import calc_correlation_matrix
+from .data_linking_functions import calc_likelihood_matrix
+from .data_linking_functions import link_prob
+from .data_linking_functions import pair_prob_approx
+from .data_linking_functions import pair_prob_hg
 
 
 SCORING_METHODS = ['metcalf', 'likescore', 'hg']
 
-from ..logconfig import LogConfig
+from ...logconfig import LogConfig
 
 
 logger = LogConfig.getLogger(__file__)
